* Added a [DynamicallyAccessedMembers] for THandler constraint in the Singleton-Class in order to remove the new() constraint which itself forces the Singleton-Handlers to have a public ctor() which is bad, and with that we avoid that.
Furthermore cleared unnessecary variables from the derived-handlers which they are not supposed to even know about, they only have to be init inside the QuestHandler.cs base class which does all the heavy inits and preloadings. it encapsulates the nessecary steps from the derived classes so that these can only focus on the handling of certain events and nothing else.