Removed a codeblock of the "Replace()" method for later rework..
Added a return value for "MoveBy()" to get pre-calculated new offset of the recently moved area, also simplified abit the "fill()" part to only clear the specific area without slicing it again..

Finally, added code to check if 2 spans are already one and same and added further code inside "Swap()" to determine if 2 spans (x, y) are indeed next to eachother. One part is finished; The first > last, now only the first < last misses.

I also moved the "first == last" out of the other if-else area because in any case whatsoever, this part is the neutral one, so once they are same in length it does not matter at all their respective offset OR if they are close or not!