Added "LengthDiff" readonly field.
Replaced the old computation of getting an address with a faster and more readable one.

Also removed public "SameMemory" field because when this bool is false, this entire struct shall throw exception cause there is no continuation when that happens..

Fixed broken logic insode the "IndexOfFirst/IndexOfLast" computation and updated "AreSameLength" logic for faster check.