Redesigned the structure of the QuestLog, such as replacing "COLOR" with Transparent, to ease the substring-replacements.
Also reworked the actual inLoop replacement of the individual parts of the questLog, which works now efficientlly. Further I use a pool to pool the created logs, in order to not unnessecary rebuild the logs from scratch over and over, whereas i can just recycle them and push them to UIRenderer.