Significantly debloated the "Comparer" classes since alot of these cheks do not have to happen on every comparison and hence got removed, now they are only done in the specific areas of need.
Renamed and reworkd "StateAndBodyComparer" to only "BodyComparer" since we only want to compare its body and the state was more or less only needed as another distinction.

Added a new Comparer: "DistanceComparer" which derives from "CellComparer" because both are in base the same but with the addition that as the name already suggests, he also compares by "Distance" not just by cell.