Removed old "Direction" enum from Grid and placed it in "Match_3.DataObjects" and redefined it to work with the "Cell" types.
Removed old "Map1DIndexAsCell()" because its not needed.

Heavily reworked "GetNextCell()" inside "WasAMatchInAnyDirection()" to respect ALL directions and store inside the "GameState" the last direction used aswell as storing the exact "First" tile which was used to even allow for a match.

Reworked also "Delete()" and added an empty "Disable()"