Removed "DoScale()" from "ICell", this is one part which will be transported to "Shape" class.
Renamed "IRectCell" to "IGridCell" because I do now distinguish between a grid which maps the internal 2Darray of tiles OR the Window-Grid which works in pixel-coordinates, which I refer to as "World" and hence I moved all the "Rectangle" code to "Shape" class.

Finally I moved the "MultiCell<TCell>" class from "MatchX" class to "Cell" file  because its a type which logically should be in there.